moving mocks closer to implementation classes.  uses barrels to aggregate mocks, service classes
