+++ app/components/clickerForm/clickerForm.ts
@@ -3,8 +3,7 @@
 import { AbstractControl, ControlGroup, FormBuilder, Validators } from '@angular/common';
 import { Component }                                              from '@angular/core';
 import { Button, Icon, Item, Label, TextInput }                   from 'ionic-angular';
-import { Clickers }                                               from '../../services/clickers';
-import { Utils }                                                  from '../../services/utils';
+import { Clickers, Utils }                                        from '../../services';
 
 @Component({
   selector: 'clicker-form',
