+++ app/services/mocks.ts
@@ -1,0 +1,1 @@
+export * from './clickers.mock';
