+++ test/diExports.ts
@@ -3,15 +3,13 @@ import { ComponentFixture, TestComponentBuilder }     from '@angular/compiler/te
 import { injectAsync }                                from '@angular/core/testing';
 import { Control }                                    from '@angular/common';
 import { App, Config, Form, NavController, Platform } from 'ionic-angular';
-import { ClickersMock, ConfigMock, NavMock }          from './mocks';
+import { ConfigMock, NavMock }                        from './mocks';
 import { Utils }                                      from '../app/services/utils';
-import { Clickers }                                   from '../app/services/clickers';
 export { TestUtils }                                  from './testUtils';
 
 export let providers: Array<any> = [
   Form,
   provide(Config, {useClass: ConfigMock}),
-  provide(Clickers, {useClass: ClickersMock}), // required by ClickerButton
   provide(App, {useClass: ConfigMock}),        // required by ClickerList
   provide(NavController, {useClass: NavMock}), // required by ClickerList
   provide(Platform, {useClass: ConfigMock}),   // -> IonicApp
